refactor: changed the structure
