+++ src/app.ts
@@ -7,11 +7,10 @@ import { router as userRouter } from './resources/users/user.router.js';
 import { router as boardRouter } from './resources/boards/board.router.js';
 import { router as taskRouter } from './resources/tasks/task.router.js';
 import { router as logEvents } from './middlewares/logging.js';
-import { router as authorization } from './middlewares/authorization.js';
+import { router as authorization } from './auth/auth.router.js';
 import { errorHandler } from './middlewares/errorHandler.js';
 import { router as validate } from './middlewares/validateUser.js';
 import { uncaughtExceptionHandler, unhandledRejectionHandler } from './middlewares/uncaughtHandler.js';
-// import { errorWrapper } from './common/errorWrapper.js';
 
 const app = express();
 const swaggerDocument = YAML.load(
