+++ src/auth/payload.model.ts
@@ -1,0 +1,8 @@
+interface IPayload {
+  id: string;
+  login: string;
+  issued?: number;
+  expires?: number;
+}
+
+export { IPayload }
